// frontend/generated/api.d.ts
export interface components {
    schemas: {
        RegisterRequest: {
            /** Format: email */
            email: string;
            password: string;
        };
        LoginRequest: {
            /** Format: email */
            email: string;
            password: string;
        };
        LoginResponse: {
            data?: {
                /** @description Temporary JWT (no tenant context) */
                access_token?: string;
                tenants?: {
                    /** Format: uuid */
                    tenant_id?: string;
                    tenant_name?: string;
                    /** @enum {string} */
                    role?: "patient" | "doctor" | "administrator";
                }[];
            };
            meta?: components["schemas"]["ResponseMeta"];
        };
// ...
        UserResponse: {
            data?: {
                /** Format: uuid */
                id?: string;
                /** Format: uuid */
                tenant_id?: string;
                /** Format: email */
                email?: string;
                /** @enum {string} */
                role?: "patient" | "doctor" | "administrator";
                /** @enum {string} */
                status?: "active" | "inactive";
                /** Format: date-time */
                created_at?: string;
            };
            meta?: components["schemas"]["ResponseMeta"];
        };
// ...
    };
// ...
}